pyproject.toml
[tool.poetry]
name = "webgen"
version = "0.1.0"
description = ""
authors = ["DrgnFireYellow <86269995+DrgnFireYellow@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
markdown = "^3.3.7"
pyyaml = "^6.0"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mkdocs = "^1.3.0"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-material = "^8.3.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
